requirements.txt
anthropic
voyageai
pandas
numpy
matplotlib
seaborn
scikit-learn
python-dotenv
langchain
langchain-community
google-generativeai
ratelimit